OMAP3: overo: Set ethaddr default to 0:0:0:0:0:0
Without this, there is a failed to set MAC address warning caused by
the absence of the variable in eth_write_hwaddr. The overo expansion
boards with ethernet all have the MAC address stored in eeprom attached
to the ethernet chip. The ethernet chip loads the address directly from
eeprom when it is reset.

Signed-off-by: Philip Balister <philip@opensdr.com>